requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pyyaml>=6.0
pdfplumber>=0.11
numpy>=1.26
nltk>=3.8
langchain-google-genai>=1.0
pytest>=8.0
httpx>=0.27
python-dotenv>=1.0
python-multipart>=0.0.9
